Mevsdb

+++ main.js
@@ -17,6 +17,18 @@ const Discord = require("discord.js"),
 	partials: ['USER','CHANNEL','GUILD_MEMBER','MESSAGE','REACTION']
 });
 
+const connection =mysql.createConnection({
+	host:cfg.dbHost,
+	user:cfg.dbLogin,
+	database:"gth_db",
+	password:cfg.dbPass
+});
+connection.query('SELECT userID from admin;',(err,data)=>{
+	data.forEach(it=>{
+		admins.push(it.userID)
+	})
+})
+
 const adminRoles ={
 	0:"sample_own",
 	1:"sample_admin",
@@ -43,6 +55,7 @@ const adminRoles ={
 	"no":true,
 };
 let own,
+	mute,
 	sample,
 	admins=[];
 
@@ -62,7 +75,7 @@ function userCommands(msg){
 			}else{
 				switch(mess[1]){
 					case"admin":
-						connectionDB(`SELECT * FROM admin WHERE uID = ${msg.author.id};`,(err,res)=>{
+						connection.query(`SELECT * FROM admin WHERE userID = ${msg.author.id};`,(err,res)=>{
 							if(res.length!==0){
 								msg.reply(
 `Here is all admins commands:
@@ -115,9 +128,9 @@ coming soon:D`
 										})
 									})
 							}else{
-								connectionDB(`SELECT * FROM admin WHERE uID = ${react.author.id};`,(err,res)=>{
+								connection.query(`select * from admin where userID = ${react.author.id};`,(err,res)=>{
 									if(err)console.log(err);
-									connectionDB(`INSERT INTO admin(trust_factor, violations) VALUES(${res.trust_factor-100}, ${res.violations+1});`,(err)=>{
+									connection.query(`insert into admin(trust_factor) values (${res.trust_factor-100});`,(err)=>{
 										if(err) return console.log(err)
 										console.log(`Фактор доверия ${react.author.id} был понижен из-за пренебрежительного отношения к боту.`)
 									})
@@ -149,27 +162,29 @@ TODO: отправка репортов с возможностью соглас
 		break;
 	}
 }
-function connectionDB(query,callback_){
-	let connection = mysql.createConnection({
-		host:cfg.dbHost,
-		user:cfg.dbLogin,
-		database:"gth_db",
-		password:cfg.dbPass
-	});
-	console.log("DB request at: "+moment().format('HH:mm:ss'))
-	connection.query(query,(err,res)=>{
-		callback_(err,res);
+function checkMuted(){
+	connection.query("SELECT * FROM mutedPPL;", (err,data)=>{
+		if(err)console.log(err)
+		data.forEach(it=>{
+			if(Number(it.endMute)<Date.now()){
+				sample.members.cache.get(it.userID).roles.remove(mute,`End of mute`)
+					.then(mbr=>{
+						if(it.roles.split("$")){
+							mbr.roles.add(it.roles.split("$"))
+						}
+						mbr.user.createDM()
+							.then(dm=>dm.send("```Your mute is over.\nAll privileges have been restored.```"))
+						connection.query("DELETE FROM mute WHERE userID=?",it.userID)
+					})
+			}
+		})
 	})
-	connection.end();
 }
 
 bot.once('ready',()=>{
-	connectionDB('SELECT uID FROM admin;', (err,data)=>{
-		data.forEach(it=>{
-			admins.push(it.uID)
-		})
-	})
-	sample=bot.guilds.cache.find(g=>g.id==='897986118077788221')//ID of guild
+	setInterval(checkMuted,5000)
+	sample=bot.guilds.cache.find(g=>g.id==='897986118077788221')//Guild
+	mute=sample.roles.cache.find(r=>r.name==="sample_muted")//Mute role
 	console.log(`${bot.user.username} is started at ${moment().format('HH:mm:ss')}`)
 });
 bot.on('messageCreate',(msg)=>{
@@ -187,22 +202,22 @@ bot.on('messageCreate',(msg)=>{
 						switch(aMess[2]){
 							case"admin":
 								if(aMess[3]){
-									connectionDB(`INSERT INTO admin(uID, perm) VALUES (${msg.mentions.members.first().id}, ${adminNums[aMess[3]]});`,(err)=>{
+									connection.query(`INSERT INTO admin(userID, perm) VALUES (${msg.mentions.members.first().id}, ${adminNums[aMess[3]]});`,(err)=>{
 										if(err)console.log(err);
 									})
 									msg.mentions.members.first().roles.add(sample.roles.cache.find(role=>role.name === adminRoles[adminNums[aMess[3]]]))
 									msg.reply("Пользователь успешно поставлен на пост!")
 								}else{
-									connectionDB("SELECT uID FROM admin WHERE perm = 'own';", (err,res)=>{
+									connection.query("SELECT userID FROM admin WHERE perm = 'own';", (err,res)=>{
 										if(err) console.log(err);
-										own=res[0].uID;
+										own=res[0].userID;
 									})
 									const col=new Discord.MessageCollector(msg.channel,n => n.author.id===own,{
 										time:3600000
 									})
 									msg.reply("own|admin|mod|obs|bot")
 									col.on("collect",n => {
-										connectionDB(`INSERT INTO admin(uID, perm) VALUES (${msg.mentions.members.first().id}, ${adminNums[n.content]});`,(err)=>{
+										connection.query(`INSERT INTO admin(userID, perm) VALUES (${msg.mentions.members.first().id}, ${adminNums[n.content]});`,(err)=>{
 											if(err)console.log(err);
 										})
 										msg.mentions.members.first().roles.add(sample.roles.cache.find(role=>role.name === adminRoles[adminNums[n.content]]))
@@ -221,13 +236,13 @@ bot.on('messageCreate',(msg)=>{
 							m.content=m.content.split(" ")
 							switch(m.content[0]){
 								case"admin":
-									connectionDB("SELECT uID FROM admin WHERE perm = 'own';", (err,res)=>{
+									connection.query("SELECT userID FROM admin WHERE perm = 'own';", (err,res)=>{
 										if(err) console.log(err);
-										own=res[0].uID;
+										own=res[0].userID;
 									})
 									if(msg.mentions.members.id){
 										if(m.content[1]){
-											connectionDB(`INSERT INTO admin(uID, perm) VALUES (${msg.mentions.members.first().id}, ${adminNums[m.content[1]]});`,(err) => {
+											connection.query(`INSERT INTO admin(userID, perm) VALUES (${msg.mentions.members.first().id}, ${adminNums[m.content[1]]});`,(err) => {
 												if(err) console.log(err);
 											})
 											msg.mentions.members.first().roles.add(sample.roles.cache.find(role => role.name===adminRoles[adminNums[m.content]]))
@@ -238,7 +253,7 @@ bot.on('messageCreate',(msg)=>{
 												time:3600000
 											})
 											col.on("collect",n => {
-												connectionDB(`INSERT INTO admin(uID, perm) VALUES (${msg.mentions.members.first().id}, ${adminNums[n.content]});`,(err) => {
+												connection.query(`INSERT INTO admin(userID, perm) VALUES (${msg.mentions.members.first().id}, ${adminNums[n.content]});`,(err) => {
 													if(err) console.log(err);
 												})
 												msg.mentions.members.first().roles.add(sample.roles.cache.find(role => role.name===adminRoles[adminNums[n.content]]))
@@ -263,20 +278,20 @@ bot.on('messageCreate',(msg)=>{
 					if(aMess[2]){
 						switch(aMess[2]){
 							case"admin":
-								connectionDB(`SELECT perm FROM admin WHERE uID = ${msg.mentions.members.first().id};`, (err,res)=>{
+								connection.query(`SELECT perm FROM admin WHERE userID = ${msg.mentions.members.first().id};`, (err,res)=>{
 									if(err) console.log(err);
 									msg.mentions.members.first().roles.remove(sample.roles.cache.find(role=>role.name === adminRoles[res[0].perm]))
 								})
-								connectionDB("DELETE FROM admin WHERE uID=?;", [msg.mentions.members.first().id],(err)=>{
+								connection.query("DELETE FROM admin WHERE userID=?;", [msg.mentions.members.first().id],(err)=>{
 									if(err)console.log(err);
 								})
 								msg.reply("Пользователь успешно убран с поста.")
 							break;
 						}
 					}else{
-						connectionDB("SELECT uID FROM admin WHERE perm = 'own';", (err,res)=>{
+						connection.query("SELECT userID FROM admin WHERE perm = 'own';", (err,res)=>{
 							if(err) console.log(err);
-							own=res[0].uID;
+							own=res[0].userID;
 						})
 						const collector=new Discord.MessageCollector(msg.channel,m => m.author.id===own,{
 							time:3600000
@@ -286,7 +301,7 @@ bot.on('messageCreate',(msg)=>{
 							switch(m.content){
 								case"admin":
 									msg.mentions.members.first().roles.remove(sample.roles.cache.find(role=>role.name === adminRoles[res[0].perm]))
-									connectionDB("DELETE FROM admin WHERE uID=?;", [msg.mentions.members.first().id],(err)=>{
+									connection.query("DELETE FROM admin WHERE userID=?;", [msg.mentions.members.first().id],(err)=>{
 										if(err)console.log(err);
 									})
 									msg.reply("Пользователь успешно убран с поста.")
@@ -309,6 +324,64 @@ bot.on('messageCreate',(msg)=>{
 					msg.channel.bulkDelete(deleteCount)
 						.catch(error => msg.reply(`Couldn't delete messages because of: ${error}`));
 				break;
+				case"!mute":
+					try{
+						let rolesIds=[];
+						if(admins.includes(msg.author.id)){
+							if (msg.mentions.users.first()) {
+								sample.members.cache.get(msg.mentions.users.first().id).roles.add(mute, `Muted by ${msg.author.tag}`)
+									.then(member_ => {//GuildMember
+										member_.roles.cache.toJSON().forEach(role=>{
+											if(role.name!=='@everyone'&&role.id!==mute.id){
+												rolesIds.push(role.id);
+												member_.roles.remove(role.id)
+											}
+										})
+										connection.query(`INSERT INTO mutedPPL(userID, endMute, roles) VALUES (${msg.mentions.members.first().id}, ${msg.content[2]?(Number(msg.content[2])*1000)+Date.now():Date.now()+86400000}, "${rolesIds.join("$")}");`,err=>{
+											if (err) console.log(err);
+										})
+									})
+							} else {
+								if (Number(msg.content[1])) {
+									sample.members.cache.get(msg.content[1]).roles.add(mute, `Muted by ${msg.author.tag}`)
+										.then(member_ => {
+											member_.roles.cache.toJSON().forEach(role=>{
+												if(role.name!=='@everyone'&&role.id!==mute.id){
+													rolesIds.push(role.id);
+													member_.roles.remove(role.id)
+												}
+											})
+											connection.query(`INSERT INTO mutedPPL(userID, endMute, roles) VALUES (${member_.id}, ${msg.content[2]?(Number(msg.content[2])*1000)+Date.now():Date.now()+86400000});`,(err) => {
+												if (err) console.log(err);
+											})
+										})
+								} else {
+									msg.reply('Вы забыли выбрать цель. Или цель указанна некорректно.')
+								}
+							}
+						}else KMS(msg);
+					}catch(err){
+						msg.reply('Something wrong... (*Window XP shutdown sound*)')
+						console.log(err)
+					}
+				break;
+				case"!getRandSong":
+					let dirInfo = fs.readdirSync("D:\\rythm_game\\osu\\Songs")
+					dirInfo.forEach((item,index)=>{
+						if(item.includes("_")){
+							dirInfo[index]=item.split("_")
+						}else{
+							dirInfo[index]=item.split(" ")
+						}
+					})
+					dirInfo.forEach((item,index)=>{
+						item.shift()
+						dirInfo[index]=item.join(" ")
+					})
+					let randNum = Math.floor(Math.random()*dirInfo.length)
+					msg.reply(dirInfo[randNum])
+					randNum=Null;
+					break;
 				default:
 					userCommands(msg)
 				break;
@@ -317,24 +390,21 @@ bot.on('messageCreate',(msg)=>{
 			userCommands(msg);
 		}
 	}else{
-		sample.members.fetch(msg.author.id)
-			.then(member=>{
-				console.log(`\x1b[35m [${moment().format("DD.MM HH:mm:ss")}]\x1b[34m\n\tusername:\x1b[0m ${msg.author.username}\x1b[34m\n\tnickname:\x1b[0m ${member.nickname}\x1b[34m\n\tuserID:\x1b[0m ${msg.author.id}\x1b[34m\n\tmsgID:\x1b[0m ${msg.id}${msg.content?'\x1b[34m\n\ttext:\x1b[0m >'+msg.content.split('\n').join(" ")+'<':''}`)// ${atch.join(",")?'\x1b[34m\n	atch:\x1b[0m '+atch.join(","):''}
-			})
-		connectionDB(`SELECT * FROM users WHERE uID = ${msg.author.id};`,(err,res)=>{
+		console.log(`${msg.author.id}-${msg.channel.name}:\n>${msg.content}<`)
+		connection.query(`SELECT * FROM users WHERE userID = ${msg.author.id};`,(err,res)=>{
 			if(err)console.log(err);
 			if(res.length!==0){
 				if(res[0].msgCount%res[0].divisor!==0){
-					connectionDB(`UPDATE users SET msgCount = ${res[0].msgCount+=1} WHERE uID = ${msg.author.id};`,(err,res)=>{
+					connection.query(`UPDATE users SET msgCount = ${res[0].msgCount+=1} WHERE userID = ${msg.author.id};`,(err)=>{
 						if(err)console.log(err);
 					})
 				}else{
-					connectionDB('UPDATE users SET msgCount=?, lvl=?, divisor=? WHERE uID=?;',[res[0].msgCount+1,res[0].lvl+1,res[0].divisor+25*(res[0].lvl+1),msg.author.id],(err,res)=>{
+					connection.query('UPDATE users SET msgCount=?, lvl=?, divisor=? WHERE userID=?;',[res[0].msgCount+1,res[0].lvl+1,res[0].divisor+25*(res[0].lvl+1),msg.author.id],(err)=>{
 						if(err) console.log(err);
 					})
 				}
 			}else{
-				connectionDB(`INSERT INTO users(uID) VALUES ('${msg.author.id}');`,(err,res)=>{
+				connection.query(`INSERT INTO users(userID) VALUES ('${msg.author.id}');`,(err)=>{
 					if(err)console.log(err);
 				})
 			}
@@ -349,10 +419,12 @@ bot.on('guildMemberAdd',mbr=>{
 	Если оценка на дне, то бот сам снимает роль. Можно также сделать такую ж автоматизацию и для юзеров, но фиг знает.
 	Дима, не забудь со всеми этими наказаниями сделать и систему повышения карьеры вплоть до модератора:D
 	*/
-	connectionDB(`SELECT * FROM users WHERE uID = '${mbr.id}';`, (err,res)=>{
+	if(mbr.bot)return;
+	connection.query(`SELECT * FROM users WHERE userID = '${mbr.id}';`, (err,res)=>{
 		if(err)console.log(err);
+		if(mbr.user.bot)return;
 		if(res.length===0){
-			connectionDB(`INSERT INTO users(uID, msgCount, lvl, banned, leaving, perm) VALUES( '${mbr.id}', 0, 0, false, 0, 0);`, (err)=>{
+			connection.query(`INSERT INTO users(userID, msgCount, lvl, banned, leaving, perm) VALUES( '${mbr.id}', 0, 0, false, 0, 0);`, (err)=>{
 				if(err)console.log(err);
 			});
 			mbr.roles.add(sample.roles.cache.get("898339903417483285"))//должна быть newbie's role
@@ -451,16 +523,31 @@ bot.on('voiceStateUpdate',(vc1,vc2)=>{
 			if(err)console.log(err)
 			if(ownID_[0])
 			if(ownID_[0]&&ownID_[0].ownID===vc1.id){
-				try{
-					sample.channels.cache.get(vc1.channelId).delete()
-						.then(()=>{
-							connectionDB('DELETE FROM voices WHERE ownID=?;',[vc1.id],(err1) => {
-								if(err1)console.log(err1)
+				if(!vc1.channel.members.size){
+					try {
+						sample.channels.cache.get(vc1.channelId).delete()
+							.then(() => {
+								connection.query('DELETE FROM voices WHERE ownID=?;', [vc1.id], err1 => {
+									if (err1) console.log(err1)
+								})
 							})
+					} catch (e) {
+						console.log(`Аэм.... Чё за? Я не могу удалити канал. памагити!!!\n ${e}`)
+						//TODO: логгирование в чат sample_logs
+					}
+				} else {
+					let nextMemberOwner = vc1.channel.members.toJSON()[Math.floor(Math.random() * (vc1.channel.members.size - 1))]
+					connection.query("UPDATE voices SET ownID=? WHERE ownID=?;",[nextMemberOwner.id, ownID_[0].ownID], (err)=>{
+						if(err)console.log(err)
+						vc1.channel.edit({name: `${nextMemberOwner.user.username}'s channel`,
+							permissionOverwrites: [
+								{
+									id: nextMemberOwner.id,
+									allow: ['MANAGE_CHANNELS', 'MANAGE_ROLES']
+								}
+							]
+						})
 					})
-				}catch(e){
-					console.log(`Аэм.... Чё за? Я не могу удалити канал. памагити!!!\n ${e}`)
-					//TODO: логгирование в чат sample_logs
 				}
 			}
 		})
@@ -468,19 +555,20 @@ bot.on('voiceStateUpdate',(vc1,vc2)=>{
 });
 bot.on('guildMemberUpdate',(oldMbr,newMbr)=>{//Сделать реагирование на покидание сервера
 	let rolesID=[];
+	if(newMbr.bot)return;
 	newMbr.roles.cache.forEach(role=>{
-		rolesID.push(role.id)
+		if(role.id!=="897986118077788221")rolesID.push(role.id);
 	})
 	if(rolesID){
-		connectionDB('UPDATE users SET roles=? WHERE uID=?;', [rolesID.join('$'), newMbr.id], err => {
+		connection.query('UPDATE users SET roles=? WHERE userID=?;', [rolesID.join('$'), newMbr.id], err => {
 			if (err) console.log(err)
 		})
 	}
 });
 bot.on('guildMemberRemove', mbr=>{
-	connectionDB('SELECT leaving FROM users WHERE uID=?;',[mbr.id],(err, result) => {
+	connection.query('SELECT leaving FROM users WHERE userID=?;',[mbr.id],(err, result) => {
 		if(err)console.log(err);
-		connectionDB('UPDATE users SET leaving=? WHERE uID=?;',[result[0].leaving+1,mbr.id],err1 => {
+		connection.query('UPDATE users SET leaving=? WHERE userID=?;',[result[0].leaving+1,mbr.id],err1 => {
 			if(err1)console.log(err1);
 		})
 	})
@@ -489,4 +577,4 @@ bot.on('guildMemberRemove', mbr=>{
 
 bot.login(cfg.token);
 
-process.on('exit',code=>console.log(`End code: `+code+` ${moment().format('HH:mm:ss')}`));
+process.on('exit',code=>console.log(`Ended with code: ${code}`));
